Move testDropSchemaExternalFiles to base class in Delta Lake
