+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/BaseDeltaLakeAwsConnectorSmokeTest.java
@@ -71,4 +71,10 @@ public abstract class BaseDeltaLakeAwsConnectorSmokeTest
                 .map(path -> format("s3://%s/%s", bucketName, path))
                 .collect(toImmutableList());
     }
+
+    @Override
+    protected String bucketUrl()
+    {
+        return format("s3://%s/", bucketName);
+    }
 }
